Handled form submission and image storing in database
