+++ package.json
@@ -12,7 +12,9 @@
     "better-sqlite3": "^9.4.3",
     "next": "14.0.3",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "slugify": "^1.6.6",
+    "xss": "^1.0.14"
   },
   "devDependencies": {
     "eslint": "^8",
